fix: allow set default container registry cleanup policy

+++ internal/sync/gitlab/types.go
@@ -5,7 +5,7 @@ type User struct {
 	isAdmin      bool
 	displayName  string
 	canCreateTLG bool
-	projectLimit int
+	projectLimit int64
 	sshKeys      []string
 }
